fix: log collector sync time cast error

+++ manager/tm/src/main/java/com/tapdata/tm/task/service/LogCollectorService.java
@@ -652,7 +652,7 @@ public class LogCollectorService {
                 return new Date();
             }
 
-            return new Date((Long) o);
+            return new Date(((Double) o).longValue());
 
         } catch (Exception e) {
             return new Date();
